finished ex 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -44,8 +44,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    //Pseudocode
+    // Find and return the data of the user who created the post
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user._id.toString() === user._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(401).json({error: 'Unauthorized'})
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
